Only calculate initial pressure norm at beginning
The SOR algorithm in "Numerical Simulation in Fluid Dynamics" uses the
initial pressure norm as one of the conditions to stop the SOR
iteration. This should be done at the very first time step, not at the
beginning of each SOR iteration.

Without this change, the pressure norm is calculated for each time step.
This can grow to be quite large, and will cause the SOR iteration to
terminate before the algorithm has a chance to converge.